+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Current Master
 
 - Integrates Rubocop linter to ensure code quality. See [#61](https://github.com/ashfurrow/danger-swiftlint/pull/61).
+- Fixes issue where `directory` variable was not escaped. See [62](https://github.com/ashfurrow/danger-swiftlint/issues/62).
 
 ## 0.10.1